Squashed 'thirdParty/mallocMC/' changes from 31f9b3c3f1..3c036261f0
3c036261f0 fix HIP support

git-subtree-dir: thirdParty/mallocMC
git-subtree-split: 3c036261f01931020998ece2f5132c39c9653ce1

+++ alpaka/include/alpaka/intrinsic/IntrinsicUniformCudaHipBuiltIn.hpp
@@ -100,7 +100,7 @@ namespace alpaka
                     std::int64_t value)
                 -> std::int32_t
                 {
-                    return __ffsll(value);
+                    return __ffsll(static_cast<unsigned long long>(value));
                 }
             };
         }
